Split PoissonRecon into pieces.
