+++ .gitignore
@@ -1,3 +1,4 @@
+*.sw*
 *.sdf
 *.suo
 /Obj/
